Moved Files into own directories
Moved Card.js and the Starwars  files into their own directory
