+++ starwars/src/components/StarWars/StarWars.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from './Card'
+import Card from '../Card/Card'
 import './StarWars.css'
